Backend/background-removal-service/requirements.txt
Flask==2.1.0
rembg==2.0.30
Pillow==8.3.2